Compilation fixes. (OpenEXR DLL flags, split of thread.h to be able to build OpenSL.)
