+++ CMakeLists.txt
@@ -437,11 +437,11 @@ if (MSVC)
     add_definitions (-D_CRT_NONSTDC_NO_WARNINGS)
     add_definitions (-D_SCL_SECURE_NO_WARNINGS)
     add_definitions (-DJAS_WIN_MSVC_BUILD)
+    add_definitions (-DOPENEXR_DLL)
     if (LINKSTATIC)
         add_definitions (-DBoost_USE_STATIC_LIBS=1)
     else ()
         add_definitions (-DBOOST_ALL_DYN_LINK)
-        add_definitions (-DOPENEXR_DLL)
     endif ()
 endif (MSVC)
